[SDK-109] Refactored the CloudCommunication Sender
