+++ njams-sdk-communication-cloud/pom.xml
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.faizsiegeln.njams</groupId>
@@ -75,6 +77,11 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
+            <groupId>com.faizsiegeln.njams</groupId>
+            <artifactId>njams4-sdk-communication-http</artifactId>
+            <version>4.0.6-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
             <scope>provided</scope>
@@ -90,8 +97,5 @@
             <artifactId>junit</artifactId>
             <scope>test</scope>
         </dependency>
-
-
-
     </dependencies>
 </project>
