+++ njams-sdk-communication-cloud/src/main/java/com/im/njams/sdk/communication/cloud/connector/CloudConnector.java
@@ -1,0 +1,4 @@
+package com.im.njams.sdk.communication.cloud.connector;
+
+public class CloudConnector {
+}
