--- njams-sdk-communication-cloud/src/main/resources/META-INF/services/com.im.njams.sdk.communication.Sender
@@ -1,1 +1,0 @@
-com.im.njams.sdk.communication.cloud.CloudSender
