+++ njams-sdk-communication-cloud/src/main/resources/META-INF/services/com.im.njams.sdk.communication.connectable.Sender
@@ -1,0 +1,1 @@
+com.im.njams.sdk.communication.cloud.connectable.CloudSender
